backend/python-ai/requirements.txt
fastapi
uvicorn
pydantic
openai
chromadb
tiktoken
python-dotenv
